feat: Table styles
This implements a .cmp-basic-table component for nice table styles. I've
gone into the partials we have to render tables and added class names so
that these styles take effect.

I'll probably want to loop back around and revisit this once we've got
our chart/table markup sorted, but this should work for now.