feat: trigger CSS animations on scroll
This introduces a new JS module: unveiler.

This JS looks for elements with classes of `.obj-unveiler` applied. When
it finds them, it watches them with an IntersectionObserver, and
triggers a class change when they are scrolled into view. This gives us
a reusable styling handle to use to apply CSS transitions and animations
when elements are scrolled into view.

This adds animation styles for the following components:

- Second set of blocks in the intro
- Emphasis highlights on headings
- Both chart types

An IntersectionObserver polyfill is provided in this PR, as current
browser support isn't great (it's currently unsupported in Safari).