feat: Redesign stacked charts
This takes into account some feedback we got on the readability of our
charts. This changes the order of the data to read more similarly to our
other chart types, reading left to right.

This also adds more definition
around exactly what "more" and "less" means in the stacked charts.

Additionally, this changes our colors for stacked charts to go dark to
light. This makes the data easier to read for some reason.