feat: Change from table/chart switch to table toggle
This changes how we're revealing tables.

Before this commit, tables replaced charts. This caused a big change in
the flow, happening before the toggle button. This resulted in a
confusing experience where it's easy to lose your place.

This commit reveals tables _after_ the button you use to toggle them
open and closed. This is a less-confusing experience, and it's more
accessible to boot. I've also added animations to help hint to the user
what's going on.