content edits to challenges
Change the section intro to “In the 2018 survey, we found that starting a design system required an executive champion and cross-departmental buy-in. In the 2019 survey, we found that adoption and maintenance were common challenges. This section explores challenges uncovered in this year’s survey.”
Change “Lack of understanding of the magnitude of a design system” to “Failure to understand the magnitude of a design system”
Remove the suffix from clear in “If hindsight is truly 20/20, planning and developing a clearer strategy early in a design system’s life could contribute to its long-term success.” so that it reads “If hindsight is truly 20/20, planning and developing a clear strategy early in a design system’s life could contribute to its long-term success.”
2nd link in “Are you planning a design system?” callout is broken. Should be https://seesparkbox.com/foundry/design_audit_for_design_system_planning 
Change the text in the above-mentioned callout from “Learn how a Discovery phase can help you identify key details to make your design system a success or how to do an audit to decomp your website and identify and organize elements for future use.” to “Learn how a Discovery phase can help you identify key details to make your design system a success. Or learn how to do an audit to decomp your website and identify and organize elements for future use.”
Change “Of the respondents who reported their design systems to be successful, 81% had a maintenance process.” to “Of the respondents who reported their design systems to be successful or very successful, 81% had a maintenance process.”
Change “Of the in-house respondents who reported their design systems to be unsuccessful, only 46% had a maintenance process.” to “Of the in-house respondents who reported their design systems to be not successful or not very successful, only 46% had a maintenance process.”


Add commas to keys of “Does your organization have a process for maintaining design system components?” chart to be “Yes, we have a process” and “No, we don’t have a process”
“Does your organization have a process for maintaining design system components?” chart should be rounded to one place

“Who approves changes to the design system?” chart should be rounded to one place 
In this chart, change “There is a single team responsible for approving changes.” to “A single team is responsible for approving changes.”
In this chart, change “There is a cross-team committee responsible for approving changes.” to “A cross-team committee is responsible for approving changes.”