Remove extra section 2 text
Section 2: The text “In house respondents to to work at larger organizations, while agency respondents work primarily at smaller companies.: should not be there

On all charts, can we get rid of the “please specify” after the “Other” like this. It will make it confusing for readers to understand. This applies to all “Other” options on every chart