Copy adjustments
On all charts, can we get rid of the “please specify” after the “Other” like this. It will make it confusing for readers to understand. This applies to all “Other” options on every chart

Section 3: Replase “As noted in 2020,” with  “As noted in the 2020 Design Systems Survey,”  and linked that way as well (This is for SEO purposes)
Section 3.1: The “Not Successful” line on this chart, the 17% on the far right should be the “I am not sure” style

Section 3.2: The “Not successful” line, the 17% should be the style of “They often or always contribute to the design system (4–5)”

Section 3.2: This quote should have some red emphasis on “concentrating on populating it”

Section 3.2: Fix Subheading text under this chart. Should read “ Answers were on a scale of 1 to 5”

Section 3.2: “MORE DEFINED PROCESS IS LINKED TO SUCCESS” chart has wrong key (Check against here)

Section 3.3: Add red emphasis to these 2 quotes:
1st quotes red “It's unavoidable”
2nd quote red “A design system will always produce some technical debt”